nixos/zsh/ohMyZsh: prioritize user package if they specify it
Currently when oh-my-zsh is enabled system-wide, and an individual user also specifies that they want to load another version of oh-my-zsh (via e.g. home-manager), only the system version gets loaded, and not the per-user-overridden version.

This is because when initializing an interactive shell, the files are loaded in the following order:
- /etc/zshenv -> we set nothing particular here
- ~/.zshenv -> here, home-manager sets `$ZSH` if package was overridden by the user
- /etc/zshrc -> we set `$ZSH` here, and run `$ZSH/oh-my-zsh.sh`
- ~/.zshrc -> here, home-manager runs `$ZSH/oh-my-zsh.sh`, sourcing again the system's `$ZSH`, not the user's `$ZSH`, because $ZSH was set by `/etc/zshrc`.

In order to allow users to override the oh-my-zsh version that gets loaded (and use the system version otherwise), I propose that we move setting the `$ZSH` variable to `/etc/zshenv`.
That way, the user can override it in their `.zshenv` before it gets sourced.
This matches what home-manager does (they set it in `~/.zshenv`) so it's probably something reasonable to do. (Although it seems [they moved it from .zshrc to here just for cleanup](https://github.com/nix-community/home-manager/commit/ee7f2413edf40f323fff76c876304ceca87931f9))
For safety, in case it's not defined (or empty) at the time of sourcing, I've opted to ask `zsh` to fallback to the value we used to give it.

A potential alternate solve would be to source the user-overridden oh-my-zsh in `.zshrc`, instead of reading `$ZSH` there (so PR home-manager to move setting `$ZSH` to `~/.zshrc`), but that seems to not be as good a solve because it implies that both versions of oh-my-zsh would actually get `source`d, and it seems unclear what side effects this may cause - especially if there are some "only execute this file once per shell pattern" laying around.